+++ TrainTicketPlatformAPI/Controllers/Admin/AdminUsersController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using TrainTicketPlatformAPI.Contracts.Admin;
 using TrainTicketPlatformAPI.Contracts.Common;
 using TrainTicketPlatformAPI.Models;
+using TrainTicketPlatformAPI.Security;
 using TrainTicketPlatformAPI.Services;
 
 namespace TrainTicketPlatformAPI.Controllers.Admin
@@ -82,6 +83,9 @@ namespace TrainTicketPlatformAPI.Controllers.Admin
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (DangerousActionGuard.RequireHeader(this, DangerousActionGuard.Delete) is { } headerError)
+                return headerError;
+
             try
             {
                 await _userService.DeleteUserAsync(id);
